vitqs_standalone/requirements.txt
# ViT-VS Standalone Requirements

# Core dependencies with CUDA support
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Computer Vision
opencv-python>=4.5.0
pillow>=8.0.0

# Scientific computing
numpy>=1.19.0
scipy>=1.6.0

# Visualization
matplotlib>=3.3.0

# Configuration
pyyaml>=5.4.0

# Additional dependencies for ViT features
transformers>=4.20.0
